// types/softcontrol/SoftControl.ts
export interface SoftControl {
    CTLID: string;          // CTLID
    CTLNAME: string;         // CTLTEXT
    CTLTEXT: string;      // CTLNAME
    // ctlType?: string;    // CTLTYPE 
    // ctlModule?: string;  // CTLMODULE
    // ctlUpd?: string;     // CTLUPD
    // costId?: string;     // COSTID
    // colId?: number;      // COLID
}